requirements.txt
requests~=2.25.1
PyJWT~=2.1.0
selenium~=3.141.0
beautifulsoup4~=4.9.3
pip~=21.1.2
setuptools~=57.0.0